# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GardenInputData = { [key: string]: unknown };

export interface GardenInput {
  data: GardenInputData;
}

export type GardenSyncPlatform = typeof GardenSyncPlatform[keyof typeof GardenSyncPlatform];


export const GardenSyncPlatform = {
  web: 'web',
  mobile: 'mobile',
} as const;

export type GardenSyncData = { [key: string]: unknown };

export interface GardenSync {
  platform: GardenSyncPlatform;
  data: GardenSyncData;
  updatedAt: string;
}

export type HabitReviewInputHabitsItem = {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @maxLength 500 */
  note?: string;
  /**
     * @minimum 1
     * @maximum 7
     */
  targetPerWeek?: number;
};

export interface HabitReviewInput {
  /**
     * @minItems 1
     * @maxItems 30
     */
  habits: HabitReviewInputHabitsItem[];
  locale?: string;
}

export interface HabitReview {
  summary: string;
  suggestions: string[];
}

export interface RevenueSummary {
  totalRevenueUsd: number;
  activeSubscriptions: number;
  updatedAt: string;
}

export interface BillingStatus {
  isPro: boolean;
}

export interface BillingRedirectInput {
  /** @pattern ^/ */
  returnPath?: string;
}

export interface BillingRedirect {
  url: string;
}