# lib/api-zod/src/generated/types/gardenInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GardenInputData } from './gardenInputData';

export interface GardenInput {
  data: GardenInputData;
}