# lib/api-zod/src/generated/types/gardenSync.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GardenSyncData } from './gardenSyncData';
import type { GardenSyncPlatform } from './gardenSyncPlatform';

export interface GardenSync {
  platform: GardenSyncPlatform;
  data: GardenSyncData;
  updatedAt: Date;
}