# lib/api-zod/src/generated/types/habitReviewInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HabitReviewInputHabitsItem } from './habitReviewInputHabitsItem';

export interface HabitReviewInput {
  /**
     * @minItems 1
     * @maxItems 30
     */
  habits: HabitReviewInputHabitsItem[];
  locale?: string;
}