# lib/api-zod/src/generated/types/revenueSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RevenueSummary {
  totalRevenueUsd: number;
  activeSubscriptions: number;
  updatedAt: Date;
}